Refactor TPI: separate handler and algorithm - Rename prop_algorithm.py to prop_algo_tpi.py and rename class PropAlgorithm to TpiAlgorithm. - Rename prop_algo_tpi.py to prop_handler_tpi.py for TPI-specific handler logic. - Update thermostat_prop.py to import handler from prop_handler_tpi.py. - Update const.py to import constants from prop_algo_tpi.py. - Update  to use TpiAlgorithm and new file structure.
- remove deprecated function_type  and constants
