+++ custom_components/versatile_thermostat/const.py
@@ -15,9 +15,7 @@ from homeassistant.components.climate.const import ClimateEntityFeature  # pylin
 from homeassistant.exceptions import HomeAssistantError
 from homeassistant.util import dt as dt_util
 
-from .prop_algorithm import (
-    PROPORTIONAL_FUNCTION_TPI,
-)
+PROPORTIONAL_FUNCTION_TPI = "tpi"
 
 from .vtherm_preset import VThermPreset, VThermPresetWithAC, VThermPresetWithAway, VThermPresetWithACAway, PRESET_TEMP_SUFFIX, PRESET_AWAY_SUFFIX  # pylint: disable=unused-import
 from .vtherm_hvac_mode import (
